# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Recetas y Usuarias API",
    "version": "1.0.0",
    "description": "Una API para la gestión de recetas y usuarios. Permite a los usuarios registrarse, iniciar sesión, ver, añadir y eliminar recetas."
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Servidor de desarrollo"
    }
  ],
  "tags": [
    {
      "name": "Recetas",
      "description": "Endpoints aplicación recetas"
    },
    {
      "name": "Usuarias"
    }
  ],
  "produces": ["application/json"],
  "paths": {
    "/api/recetas": {
      "get": {
        "tags": ["Recetas"],
        "summary": "Lista todas las recetas"
      },
      "post": {
        "tags": ["Recetas"],
        "summary": "Crea una nueva receta",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/recetas"
              }
            }
          }
        }
      }
    },
    "/api/recetas/{id}": {
      "get": {
        "tags": ["Recetas"],
        "summary": "Obtiene detalles de una receta específica",
        "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Muestra una receta por su ID",
              "required": true,
              "type": "string",
              "schema": {
                "type": "integer"
              }
            }
        ]
      },
      "put": {
        "tags": ["Recetas"],
        "summary": "Actualiza una receta específica",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Filtra las recetas por su ID",
            "required": true,
            "type": "string",
            "schema": {
              "type": "integer"
            }
          }
        ]
      },
      "delete": {
        "tags": ["Recetas"],
        "summary": "Elimina una receta específica",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Elimina una receta por su ID",
            "required": true,
            "type": "string",
            "schema": {
              "type": "integer"
            }
          }
        ]
      }
    },
    "/registro": {
      "post": {
        "tags": ["Usuarias"],
        "description": "Registra una usuaria",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/usuarios"
              }
            }
          }
        }
      }
    },
    "/login": {
      "post": {
        "tags": ["Usuarias"],
        "description": "Permite que una usuaria inicie sesión",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/usuarios"
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "recetas": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "El identificador único de la receta en la base de datos."
          },
          "nombre": {
            "type": "string",
            "description": "El nombre de la receta."
          },
          "ingredientes": {
            "type": "string",
            "description": "Una lista de ingredientes necesarios para la receta."
          },
          "instrucciones": {
            "type": "string",
            "description": "Las instrucciones paso a paso para preparar la receta."
          }
        }
      },
      "usuarios": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "El identificador único de usuario en la base de datos."
          },
          "nombre": {
            "type": "string",
            "description": "Nombre de usuario."
          },
          "email": {
            "type": "string",
            "description": "Email del usuario"
          },
          "password": {
            "type": "string",
            "description": "Contraseña del usuario."
          }
        }
      }
    }
  }
}